26/27 launch: live-season kickoff fixes
- schemas: teams[].strength nullable (FPL seeds it later in pre-season)
- cli: kickoff/report called sibling typer commands as plain functions,
  leaving unpassed bool options as truthy OptionInfo — kickoff's refresh
  ran in --raw mode and never ingested the live season, so the whole
  chain silently ran on stale provisional data; pass raw/force explicitly
  (+ regression test)
- dataset/horizon/features: flatten pandas masked-int dtypes from NULL
  strength columns to float64 — LightGBM rejects extension dtypes
- retire the pre-launch provisional machinery (deploy step, script,
  fixture JSON, 'provisional' publish state + site branches): real
  pre-GW1 live data also has zero played rows, so the detector would
  have mislabeled the live draft

+++ src/fplscout/models/horizon.py
@@ -143,7 +143,11 @@ def build_horizon_ev(
 
     team_fixtures = _team_fixtures_long(fixtures)
     id_to_code = dict(zip(teams["team_id"], teams["code"], strict=False))
-    strength_by_team = dict(zip(teams["team_id"], teams["strength"], strict=False))
+    # astype first: strength is NULL pre-season (masked Int32, pd.NA values);
+    # Series-level astype maps NA→NaN, so the dict holds plain floats LightGBM eats
+    strength_by_team = dict(
+        zip(teams["team_id"], teams["strength"].astype("float64"), strict=False)
+    )
 
     # Drop every column that's specific to the DECISION gw's own fixture — all of
     # these get replaced per target gw by the merge below; keeping them around
